房间信息接口，已测试

+++ backend/src/main/scala/com/neo/sk/tank/protocol/WatchGameProtocol.scala
@@ -7,15 +7,15 @@ object WatchGameProtocol {
   //esheep请求房间号
   case class GetRoomIdReq(playerId:Long)
   case class GetRoomId(playerId:Long,replyTo:ActorRef[RoomIdRsp]) extends RoomManager.Command
-  case class RoomInfo(roomId:Int)
+  case class RoomInfo(roomId:Long)
   case class RoomIdRsp(
                         data:RoomInfo,
                         errCode:Int = 0,
                         msg:String = "ok")
 
   //根据房间Id获取该房间的用户列表
-  case class GetUserInfoListReq(roomId:Int)
-  case class GetUserInfoList(roomId:Int,replyTo:ActorRef[UserInfoListByRoomIdRsp]) extends RoomManager.Command
+  case class GetUserInfoListReq(roomId:Long)
+  case class GetUserInfoList(roomId:Long,replyTo:ActorRef[UserInfoListByRoomIdRsp]) extends RoomManager.Command
   case class UserInfo(
                        uId:Long,
                        name:String
@@ -29,7 +29,7 @@ object WatchGameProtocol {
 
   //获取进行中游戏房间列表
   case class GetRoomListReq(replyTo:ActorRef[RoomListRsp]) extends RoomManager.Command
-  case class RoomList(roomList:List[Int])
+  case class RoomList(roomList:List[Long])
   case class RoomListRsp(data:RoomList,
                          errCode:Int = 0,
                          msg:String = "ok")
